:recycle: Refactor: presentActivityView 메서드를 PresentableActivityView 프로토콜로 분리
